Fix integer overflow at ~53M panels by converting to long
## Description
Fixes an integer overflow issue that occurred at approximately 53,000,000 panels.

## Why?
Players should be able to increase the panel limit without worrying about overflows :)

## Changes
- Changed panel count from `int` to `long` to support higher values
- Updated related calculations and operations to handle `long` values

## Testing
- Basic verification: Set panel count to 500M and confirmed energy transfer works without crashes

## Type of Change
- [x] Bug fix (non-breaking change which fixes an issue)

+++ src/main/java/com/buuz135/dysoncubeproject/client/gui/DysonProgressGuiAddon.java
@@ -44,7 +44,7 @@ public class DysonProgressGuiAddon extends BasicScreenAddon {
         ++y;
         guiGraphics.drawString(Minecraft.getInstance().font, Component.translatable("gui.dysoncubeproject.progress", new DecimalFormat().format(dyson.getProgress() * 100)).withStyle(ChatFormatting.BLUE), this.getPosX() + guiX, this.getPosY() + guiY + Minecraft.getInstance().font.lineHeight * y, 0xFFFFFF, false);
         ++y;
-        guiGraphics.drawString(Minecraft.getInstance().font, Component.translatable("gui.dysoncubeproject.power_gen", NumberUtils.getFormatedBigNumber(dyson.getSolarPanels() * Config.POWER_PER_SAIL)).withStyle(ChatFormatting.BLUE), this.getPosX() + guiX, this.getPosY() + guiY + Minecraft.getInstance().font.lineHeight * y, 0xFFFFFF, false);
+        guiGraphics.drawString(Minecraft.getInstance().font, Component.translatable("gui.dysoncubeproject.power_gen", NumberUtils.getFormatedBigNumber((long) dyson.getSolarPanels() * Config.POWER_PER_SAIL)).withStyle(ChatFormatting.BLUE), this.getPosX() + guiX, this.getPosY() + guiY + Minecraft.getInstance().font.lineHeight * y, 0xFFFFFF, false);
         ++y;
         guiGraphics.drawString(Minecraft.getInstance().font, Component.translatable("gui.dysoncubeproject.power_con", NumberUtils.getFormatedBigNumber(dyson.getLastConsumedPower())).withStyle(ChatFormatting.BLUE), this.getPosX() + guiX, this.getPosY() + guiY + Minecraft.getInstance().font.lineHeight * y, 0xFFFFFF, false);
         ++y;
